fixed folder structure for arm files, not commands
