+++ Team-3749-2025/src/main/java/frc/robot/Robot.java
@@ -10,9 +10,9 @@ import edu.wpi.first.wpilibj.TimedRobot;
 import edu.wpi.first.wpilibj.simulation.DriverStationSim;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.CommandScheduler;
-import frc.robot.subsystems.arm.armJavaFiles.AlgaeArm;
-import frc.robot.subsystems.arm.armJavaFiles.ClimbArm;
-import frc.robot.subsystems.arm.armJavaFiles.CoralArm;
+import frc.robot.subsystems.arm.AlgaeArm;
+import frc.robot.subsystems.arm.ClimbArm;
+import frc.robot.subsystems.arm.CoralArm;
 import frc.robot.subsystems.example.ExampleSubsystem;
 import frc.robot.subsystems.swerve.Swerve;
 import frc.robot.utils.ShuffleData;
